tests/wpt/tests/notifications/cross-origin-serviceworker.tentative.https.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notifications in cross origin iframes</title>
<link rel="help" href="https://github.com/whatwg/notifications/issues/177">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/helpers.js"></script>
<script>
// The syntax below will give us a third party URL.
// See: https://web-platform-tests.org/writing-tests/server-features.html
const thirdPartyOrigin = 'https://{{hosts[alt][]}}:{{ports[https][0]}}';
const thirdPartyIframe =
  `${thirdPartyOrigin}/notifications/resources/cross-origin-serviceworker-iframe.sub.html`;
let promise;

promise_setup(async () => {
  await trySettingPermission("granted");

  promise = new Promise(r => window.addEventListener("message", ev => {
    if ("shown" in ev.data) {
      r(ev.data)
    }
  }));

  const iframe = document.createElement("iframe");
  iframe.src = thirdPartyIframe;
  document.body.append(iframe);
})

promise_test(async t => {
  const result = await promise;
  assert_false(result.shown, `notification should not be shown`);
}, "third party serviceworker");
</script>
